### 2024-05-28 #6795 @editorialbot
[PRE REVIEW]: VOTCA: multiscale frameworks for quantum and classical simulations in soft matter
**Submitting author:** @junghans (<a href="http://orcid.org/0000-0003-0925-1458">Christoph Junghans</a>)
**Repository:** https://github.com/votca/votca
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v2024.1
