+++ share/sphinx/xtp/bibliography.rst
@@ -1,0 +1,5 @@
+Bibliography
+============
+
+.. [Ruehle:2009.a] V. Rühle, C. Junghans, A. Lukyanov, K. Kremer, and D. Andrienko, Versatile Object-oriented Toolkit for Coarse-graining Applications, J. Chem. Theor. Comp. 5, 3211 (2009), https://dx.doi.org/10.1021/ct900369w
+.. [Neese:2023] F. Neese, The ORCA Program System, WIREs Computational Molecular Science 1, 73 (2012) 
