+++ share/sphinx/xtp/XTP-MANUAL.rst
@@ -2,9 +2,46 @@
 XTP Manual
 **********
 
+**Disclaimer:**
+This manual is not complete. The best way to start using the software is
+to look at the provided tutorials. The reference section is generated
+automatically from the source code of the master version, so please make sure that your
+software and manual versions match.
+
+
 .. toctree::
    xtp_introduction
    theory
    pyxtp
    workflows_and_calculators
    reference
+   bibliography
+
+
+Citations
+---------
+
+Development of this software depends on academic research grants. If
+you are using the package, please cite the following papers:
+
+- [scherer_understanding_2018]_ Force-matching with three-body interactions in VOTCA
+- [mashayakrelative]_ Relative entropy and optimization-driven coarse-graining methods in VOTCA
+- [ruhle2011hybrid]_ Hybrid approaches to coarse-graining using the VOTCA package: liquid hexane
+- [Ruehle:2009.a]_ Versatile Object-oriented Toolkit for Coarse-graining Applications
+
+
+Development
+-----------
+
+The core development is currently taking place at the Eindhoven University of Technology, Netherlands, and Los Alamos National Laboratory.
+
+
+Copyright
+---------
+
+VOTCA is free software. The entire package is available under the Apache
+License. For details, check the LICENSE file in the source code. The
+VOTCA source code is available on our homepage, .
+
+
+
