### 2024-05-28 0d0d69c
GW part
