### 2024-05-28 bf0e790
BSE part
