+++ share/sphinx/xtp/theory.rst
@@ -23,7 +23,7 @@ Using the ground-state reference, many-body Green's functions theory with the :m
 
     \left\{ -\frac{\hbar^2}{2m}\nabla^2 + V_\textrm{ext}(\mathbf{r}) + V_\textrm{H}(\mathbf{r})\right\}\phi_n^\textrm{QP}(\mathbf{r}) + \int{\Sigma(\mathbf{r},\mathbf{r}',\varepsilon_n^\textrm{QP})\phi_n^\textrm{QP}(\mathbf{r}')d\mathbf{r}'} = \varepsilon_n^\textrm{QP} \phi_n^\textrm{QP}(\mathbf{r}) .
 
-In place of the exchange-correlation potential in Eq.:eq:`equ:KS`, the energy-dependent self-energy operator :math:`\Sigma(\mathbf{r},\mathbf{r}',E) occurs in the QP equations. This operator is evaluated using the one-body Green's function in quasi-particle approximation
+In place of the exchange-correlation potential in Eq.:eq:`equ:KS`, the energy-dependent self-energy operator :math:`\Sigma(\mathbf{r},\mathbf{r}',E)` occurs in the QP equations. This operator is evaluated using the one-body Green's function in quasi-particle approximation
 
 .. math::
     :label: equ:Green
@@ -46,6 +46,57 @@ where :math:`W` denotes the dynamically screened Coulomb interaction. Assuming t
 
 
 As the correction :math:`\Delta \varepsilon_n^{GW}` itself depends on :math:`\varepsilon_n^\textrm{QP}`, Eq.:eq:`equ:theory:gw_sc` needs to be solved self-consistently.
+
+Neutral excitations with a conserved number of electrons can be obtained from the Bethe-Salpeter Equation (BSE) by expressing coupled electron-hole amplitudes of excitation :math:`S` in a product basis of single-particle orbitals, i.e., 
+
+.. math:: 
+    :label: equ:bsewf
+    
+    \chi_S(\mathbf{r}_\textrm{e},\mathbf{r}_\textrm{h})=\sum_{v}^{\mathrm{occ}}\sum_c^{\mathrm{unocc}}A_{vc}^S\phi_{c}(\mathbf{r}_\textrm{e})\phi^*_{v}(\mathbf{r}_\textrm{h})+B_{vc}^S\phi_{v}(\mathbf{r}_\textrm{e})\phi^{*}_{c}(\mathbf{r}_\textrm{h}),
+  
+where :math:`\mathbf{r}_\textrm{e}` (:math:`\mathbf{r}_\textrm{h}`) is for the electron (hole) coordinate and :math:`A_{vc}` (:math:`B_{vc}`) are the expansion coefficients of the excited state wave function in terms of resonant (anti-resonant) transitions between occupied :math:`v` and unoccupied :math:`c` states, respectively. In this basis, the BSE turns into an effective two-particle Hamiltonian problem of the form
+
+.. math::
+    :label: equ:theory:bseeigenvalue
+
+    \begin{pmatrix}
+                                    \underline{\mathbf{H}}^{\text{res}}&\underline{\mathbf{K}} \\
+                                    -\underline{\mathbf{K}} & -\underline{\mathbf{H}}^{\text{res}}
+                                   \end{pmatrix}
+ \begin{pmatrix}
+ \mathbf{A}^S\\ \mathbf{B}^S\
+ \end{pmatrix}
+=\Omega_S
+\begin{pmatrix}
[…]
