+++ share/sphinx/xtp/theory.rst
@@ -4,7 +4,7 @@ Theory and Background
 Density-Functional Theory 
 *************************
 
-Excited state calculations require a reference ground state calculation within density-functional theory. VOTCA-XTP provides both an automated interface to the ORCA package [Neese:2012]_ and a lightweight internal DFT engine based on atom-centered Gaussian-type orbitals for method developing and testing. It solves the Kohn-Sham Equations for the molecular orbitals :math:`\phi_n^\text{KS}(\mathbf{r})` with orbital energies :math:`\varepsilon_n^\textrm{KS}$`
+Excited state calculations require a reference ground state calculation within density-functional theory. VOTCA-XTP provides both an automated interface to the ORCA package [Neese:2012]_ and a lightweight internal DFT engine based on atom-centered Gaussian-type orbitals for method developing and testing. It solves the Kohn-Sham Equations for the molecular orbitals :math:`\phi_n^\text{KS}(\mathbf{r})` with orbital energies :math:`\varepsilon_n^\text{KS}$`
 
 .. math::
     :label: equ:KS
