### 2024-05-28 b46ac53
Update papers/v2024/paper.md
Co-authored-by: Jakub Krajniak <563684+jkrajniak@users.noreply.github.com>
