### 2024-05-28 954ba7b
update CHANGELOG.rst
