### 2024-05-28 #6795 @Kevin-Mattheus-Moerman
@junghans I fixed the title as requested. I've also just invited a potential handling editor.

For the moment, could you add contributing guidelines to your project, and link to them in the README? They appear missing (examples here: https://contributing.md/example/).
