+++ papers/v2024/paper.bib
@@ -425,7 +425,7 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 }
 
 
-@software{mpi-votca,
+@software{mpip-votca,
 	title = {MPI-P Fork of VOTCA},
 	url = {https://gitlab.mpcdf.mpg.de/votca/votca}
 }
