### 2024-05-29 #6795 @junghans
> For the moment, could you add contributing guidelines to your project, and link to them in the README? They appear missing (examples here: https://contributing.md/example/).

@Kevin-Mattheus-Moerman this is done.
