+++ CHANGELOG.rst
@@ -22,8 +22,8 @@ Version 2024-dev
 -  add CI on FreeBSD (#1124)
 -  Fix Sphinx build (#1131)
 -  add CONTRIBUTING.md (#1133)
--  Added v2024 JOSS paper and link to older pubs (#1132)
 -  csg_table fix (#1134)
+-  Added v2024 JOSS paper and link to older pubs (#1132)
 
 Version 2024 (released 22.01.24)
 ================================
