### 2024-05-30 1d0c269
paper.md: added LANL aknowl.
