+++ papers/make_paper.sh
@@ -1,1 +1,2 @@
-docker run --rm  --volume $PWD/v2024:/data  --user $(id -u):$(id -g)  --env JOURNAL=joss  openjournals/paperdraft
+#! /bin/bash
+docker run --rm  --volume "$PWD/v2024:/data"  --user "${UID}:${GROUPS[0]}"  --env JOURNAL=joss  openjournals/paperdraft
