+++ csg/src/libcsg/version.cc
@@ -43,7 +43,7 @@ void HelpTextHeader(const std::string &tool_name) {
   std::cout << "==================================================\n"
             << "========   VOTCA (http://www.votca.org)   ========\n"
             << "==================================================\n\n"
-            << "please read and cite: "PROJECT_CITATION "\n"
+            << "please read and cite: " PROJECT_CITATION "\n"
             << "please submit bugs to " PACKAGE_BUGREPORT "\n\n"
             << tool_name << ", version " << votca::tools::ToolsVersionStr()
 #ifdef GMX_DOUBLE
