### 2024-05-31 #6795 @junghans
@srmnitc please let us know if you need anything else from us. And btw @VOTCA originated at the MPI for Polymer Research.
