### 2024-06-03 #6795 @srmnitc
👋🏽 @TariniHardikar @robashaw  @apoma @naik-aakash would any of you be willing to  review this submission? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).
