### 2024-06-04 #6795 @srmnitc
👋🏽 @gomartini-collab would any of you be willing to review this submission? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).
