### 2024-06-05 #6795 @srmnitc
👋🏽 @jngrad @susilehtola @HugoStrand would you be willing to review this submission? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).
