### 2024-06-07 b187e9c
Format code using clang-format version 18.1.6 (Fedora 18.1.6-3.fc40)
