+++ xtp/include/votca/xtp/region.h
@@ -47,7 +47,7 @@ class StaticRegion;
 class Region {
 
  public:
-  Region(Index id, Logger& log) : id_(id), log_(log){};
+  Region(Index id, Logger& log) : id_(id), log_(log) {};
   virtual ~Region() = default;
 
   virtual void WriteToCpt(CheckpointWriter& w) const = 0;
