+++ xtp/src/libxtp/dftengine/embeddingengine.cc
@@ -452,10 +452,11 @@ bool DFTEngine::EvaluateTruncatedActiveRegion(Orbitals& trunc_orb) {
           << TimeStamp() << " E_trunc (Ha) = "
           << TruncatedDensityMatrix.cwiseProduct(H0_trunc_).sum() +
                  E_Hartree_truncated + E_xc_truncated
-          << "\n \t \t \t" << " E_fullDFT (Ha) = " << Total_E_full_
           << "\n \t \t \t"
+          << " E_fullDFT (Ha) = " << Total_E_full_ << "\n \t \t \t"
           << " E_initial_trunc(Ha) = " << Initial_truncated_energy
-          << "\n \t \t \t" << " E_embedding_correction(Ha) = "
+          << "\n \t \t \t"
+          << " E_embedding_correction(Ha) = "
           << ((TruncatedDensityMatrix - InitialActiveDmat_trunc_)
                   .cwiseProduct(v_embedding_trunc_)
                   .sum())
@@ -463,7 +464,8 @@ bool DFTEngine::EvaluateTruncatedActiveRegion(Orbitals& trunc_orb) {
 
       XTP_LOG(Log::error, *pLog_)
           << " Truncated Energy of the system: E_trunc + E_fullDFT - "
-          << "\n \t \t" << " E_initial_trunc + E_embedding_correction = "
+          << "\n \t \t"
+          << " E_initial_trunc + E_embedding_correction = "
           << TruncatedTotalEnergy << std::flush;
 
       PrintMOs(MOs_trunc.eigenvalues(), Log::info);
