+++ papers/v2024/paper.bib
@@ -79,7 +79,7 @@ volume = {208},
 pages = {149-161},
 year = {2016},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2016.06.019},
+doi = {10.1016/j.cpc.2016.06.019},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465516301990},
 author = {Fabien Bruneval and Tonatiuh Rangel and Samia M. Hamed and Meiyue Shao and Chao Yang and Jeffrey B. Neaton},
 keywords = {Electronic structure of molecules, Many-body perturbation theory,  approximation, Bethe–Salpeter equation},
@@ -89,13 +89,14 @@ abstract = {We summarize the molgw code that implements density-functional theor
 
 
 
- @article{manby2012simple,
+@article{manby2012simple,
   title={A Simple, Exact Density-Functional-Theory Embedding Scheme},
   author={Manby, F. R. and Stella, M. and Goodpaster, J. D. and Miller, T. F.},
   journal={J. Chem. Theory Comput.},
   year={2012},
   volume={8},
-  pages={2564--2568}
+  pages={2564--2568},
+  doi={10.1021/ct300544e}
 }
 
 @book{intelMKL,
@@ -340,7 +341,7 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 	pages = {1546-1553},
 	year = {2014},
 	issn = {0010-4655},
-	doi = {https://doi.org/10.1016/j.cpc.2014.01.018},
+	doi = {10.1016/j.cpc.2014.01.018},
 	url = {https://www.sciencedirect.com/science/article/pii/S0010465514000447},
 	author = {Pierre {de Buyl} and Peter H. Colberg and Felix Höfling},
 	keywords = {Molecular simulation, HDF5},
@@ -390,7 +391,7 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 	number = {7},
 	pages = {1475-1487},
 	keywords = {molecular dynamics simulation package, coarse-graining simulation, iterative Boltzmann inversion method},
-	doi = {https://doi.org/10.1002/jcc.21717},
+	doi = {10.1002/jcc.21717},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21717},
 	eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.21717},
 	year = {2011}
