### 2024-06-10 80b2cdb
add some more DOIs
