+++ papers/v2024/paper.bib
@@ -99,11 +99,17 @@ abstract = {We summarize the molgw code that implements density-functional theor
   doi={10.1021/ct300544e}
 }
 
-@book{intelMKL,
-    title = {{Intel Math Kernel Library. Reference Manual}},
-    year = {2009},
-    publisher = {Intel Corporation},
-    isbn = {630813-054US}
+@inbook{intelMKL,
+	author="Wang, Endong and Zhang, Qing and Shen, Bo and Zhang, Guangyong and Lu, Xiaowei and Wu, Qing and Wang, Yajuan",
+	title="Intel Math Kernel Library",
+	bookTitle="High-Performance Computing on the Intel® Xeon Phi{\texttrademark}: How to Fully Exploit MIC Architectures",
+	year="2014",
+	publisher="Springer International Publishing",
+	address="Cham",
+	pages="167--188",
+	isbn="978-3-319-06486-4",
+	doi="10.1007/978-3-319-06486-4_7",
+	url="https://doi.org/10.1007/978-3-319-06486-4_7"
 }
 
 @article{Shaw2017,
@@ -329,6 +335,7 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
   number={27},
   pages={273002},
   url={http://stacks.iop.org/0953-8984/29/i=27/a=273002},
+  doi={10.1088/1361-648X/aa680e},
   year={2017},
   abstract={The atomic simulation environment (ASE) is a software package written in the Python programming language with the aim of setting up, steering, and analyzing atomistic simulations. In ASE, tasks are fully scripted in Python. The powerful syntax of Python combined with the NumPy array library make it possible to perform very complex simulation tasks. For example, a sequence of calculations may be performed with the use of a simple ‘for-loop’ construction. Calculations of energy, forces, stresses and other quantities are performed through interfaces to many external electronic structure codes or force fields using a uniform interface. On top of this calculator interface, ASE provides modules for performing many standard simulation tasks such as structure optimization, molecular dynamics, handling of constraints and performing nudged elastic band calculations.}
 }
@@ -347,12 +354,18 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 	keywords = {Molecular simulation, HDF5},
 }
 
-@software{hdf5,
-	author = {{The HDF Group}},
-	title = {{Hierarchical Data Format, version 5}},
-	url = {https://github.com/HDFGroup/hdf5}
+@inbook{hdf5,
+	author="Koranne, Sandeep",
+	title="Hierarchical Data Format 5 : HDF5",
+	bookTitle="Handbook of Open Source Tools",
+	year="2011",
+	publisher="Springer US",
+	address="Boston, MA",
+	pages="191--200",
+	isbn="978-1-4419-7719-9",
+	doi="10.1007/978-1-4419-7719-9_10",
+	url="https://doi.org/10.1007/978-1-4419-7719-9_10"
 }
-
 
 @article{bocs,
 	author = {Dunn, Nicholas J. H. and Lebold, Kathryn M. and DeLyser, Michael R. and Rudzinski, Joseph F. and Noid, W.G.},
