+++ papers/v2024/paper.bib
@@ -12,15 +12,6 @@ ceton University Press},
      url = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
  }
 
- @misc{fidgit,
-   author = {A. M. Smith and K. Thaney and M. Hahnel},
-   title = {Fidgit: An ungodly union of GitHub and Figshare},
-   year = {2020},
-   publisher = {GitHub},
-   journal = {GitHub repository},
-   url = {https://github.com/arfon/fidgit}
- }
-
  @article{pyscf,
     author = {Sun, Qiming and Zhang, Xing and Banerjee, Samragni and Bao, Peng and Barbry, Marc and Blunt, Nick S. and Bogdanov, Nikolay A. and Booth, George H. and Chen, Jia and Cui, Zhi-Hao and Eriksen, Janus J. and Gao, Yang and Guo, Sheng and Hermann, Jan and Hermes, Matthew R. and Koh, Kevin and Koval, Peter and Lehtola, Susi and Li, Zhendong and Liu, Junzi and Mardirossian, Narbe and McClain, James D. and Motta, Mario and Mussard, Bastien and Pham, Hung Q. and Pulkin, Artem and Purwanto, Wirawan and Robinson, Paul J. and Ronca, Enrico and Sayfutyarova, Elvira R. and Scheurer, Maximilian and Schurkus, Henry F. and Smith, James E. T. and Sun, Chong and Sun, Shi-Ning and Upadhyay, Shiv and Wagner, Lucas K. and Wang, Xiao and White, Alec and Whitfield, James Daniel and Williamson, Mark J. and Wouters, Sebastian and Yang, Jun and Yu, Jason M. and Zhu, Tianyu and Berkelbach, Timothy C. and Sharma, Sandeep and Sokolov, Alexander Yu. and Chan, Garnet Kin-Lic},
     title = "{Recent developments in the PySCF program package}",
