### 2024-06-10 #6864 @junghans
> - No DOI given, and none found for title: Libint: A library for the evaluation of molecular ...
> - No DOI given, and none found for title: Eigen v3
> - No DOI given, and none found for title: pybind11 – Seamless operability between C++11 and ...

We cited those 3 as suggested in their documents, no DOI available.

> - No DOI given, and none found for title: MPI-P Fork of VOTCA

This one has no DOI either, and the repo is also currently unavailable, still investigating, but if that persists we might drop this citation.
