### 2024-06-10 #6864 @srmnitc
@junghans thanks! My recommendation would be add the `url` field in the software citations which would ensure that link appears in the paper. Without a direct link in the paper, it could be hard to find.
