### 2024-06-11 51c5bd5
Update CHANGELOG.rst
