+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ Version 2024-dev
 -  DFT and GW-BSE theory section in manual (#1135)
 -  add rate theory part to XTP manual section (#1136)
 -  adding transport model section to XTP manual (#1137)
+-  switching optimizer in XTP tutorial (#1139, #1140)
 
 Version 2024 (released 22.01.24)
 ================================
