### 2024-06-11 cc8752e
Add citation of md software packages
