+++ papers/v2024/paper.bib
@@ -472,3 +472,42 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 	author = {Stuart Slattery and Samuel Temple Reeve and Christoph Junghans and Damien Lebrun-Grandié and Robert Bird and Guangye Chen and Shane Fogerty and Yuxing Qiu and Stephan Schulz and Aaron Scheinberg and Austin Isner and Kwitae Chong and Stan Moore and Timothy Germann and James Belak and Susan Mniszewski},
 	title = {Cabana: A Performance Portable Library for Particle-Based Simulations}, journal = {Journal of Open Source Software}
 }
+
+@article{LAMMPS,
+  author = "A. P. Thompson and H. M. Aktulga and R. Berger and 
+     D. S. Bolintineanu and W. M. Brown and P. S. Crozier and
+     P. J. in 't Veld and A. Kohlmeyer and S. G. Moore and T. D. Nguyen and
+     R. Shan and M. J. Stevens and J. Tranchida and C. Trott and S. J. Plimpton",
+  title = "{LAMMPS} - a flexible simulation tool for
+     particle-based materials modeling at the 
+     atomic, meso, and continuum scales",
+  journal = "Comp. Phys. Comm.",
+  volume =  "271",
+  pages =   "108171",
+  year =    "2022",
+  doi = "10.1016/j.cpc.2021.108171"
+} 
+
+@article{GUZMAN201966,
+title = {ESPResSo++ 2.0: Advanced methods for multiscale molecular simulation},
+journal = {Computer Physics Communications},
+volume = {238},
+pages = {66-76},
+year = {2019},
+issn = {0010-4655},
+doi = {10.1016/j.cpc.2018.12.017},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465518304399},
+author = {Horacio V. Guzman and Nikita Tretyakov and Hideki Kobayashi and Aoife C. Fogarty and Karsten Kreis and Jakub Krajniak and Christoph Junghans and Kurt Kremer and Torsten Stuehn},
+}
+
+
+@article{weik2019espresso,
+  title={ESPResSo 4.0--an extensible software package for simulating soft matter systems},
+  author={Weik, Florian and Weeber, Rudolf and Szuttor, Kai and Breitsprecher, Konrad and de Graaf, Joost and Kuron, Michael and Landsgesell, Jonas and Menke, Henri and Sean, David and Holm, Christian},
+  journal={The European Physical Journal Special Topics},
+  volume={227},
+  pages={1789--1816},
+  year={2019},
+  publisher={Springer},
+  doi={10.1140/epjst/e2019-800186-9}
+}
