### 2024-06-11 c2de6cf
paper.bib: more DOIs
