+++ papers/v2024/paper.bib
@@ -121,6 +121,7 @@ year = {2017}
   author = "E.~F.~Valeev",
   title = "Libint: A library for the evaluation of molecular integrals of many-body operators over Gaussian functions",
   howpublished = "http://libint.valeyev.net/",
+  doi = {10.5281/zenodo.10780719},
   note = "version 2.9.0",
   year =         2024
 }
@@ -433,6 +434,9 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 
 @software{mpip-votca,
 	title = {MPI-P Fork of VOTCA},
+        author = {Andrienko, Denis and {others}},
+        version = {2023.1},
+        date = {2023-12-07},
 	url = {https://gitlab.mpcdf.mpg.de/votca/votca}
 }
 
