### 2024-06-11 #6864 @junghans
> > * No DOI given, and none found for title: MPI-P Fork of VOTCA
>
> This one has no DOI either, and the repo is also currently unavailable, still investigating, but if that persists we might drop this citation.

The repo is back online.
