### 2024-06-13 88c87bd
csg-tutorials: auto-detect gmx binary
