+++ csg-tutorials/urea-water/atomistic/run.sh
@@ -2,9 +2,11 @@
 
 echo "Start simulaion"
 
-gmx grompp -v
+gmx=$(type -p gmx || type -p gmx_d || { echo "gmx not found" >&2; exit 1; })
 
-gmx mdrun -v
+$gmx grompp -v
+
+$gmx mdrun -v
 
 echo "Finished simulation"
 
