### 2024-06-13 #6864 @srmnitc
> **This is a bit ambiguous**: Human and animal research: If the paper contains original data research on humans subjects or animals, does it comply with JOSS's human participants research policy and/or animal research policy? If the paper contains no such data, please check this item. So what to do: check if it is affirmative and check if it not affirmative? @srmnitc

In this case, since there is on human/animal research, checking the box would be the appropriate response.
