+++ papers/make_paper.sh
@@ -1,2 +1,2 @@
 #! /bin/bash
-docker run --rm  --volume "$PWD/v2024:/data"  --user "${UID}:${GROUPS[0]}"  --env JOURNAL=joss  openjournals/paperdraft
+docker run --rm  --volume "$PWD/v2024.1:/data"  --user "${UID}:${GROUPS[0]}"  --env JOURNAL=joss  openjournals/paperdraft
