+++ .github/workflows/draft-pdf.yml
@@ -15,7 +15,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: papers/v2024/paper.md
+          paper-path: papers/v2024.1/paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -23,4 +23,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: papers/v2024/paper.pdf
+          path: papers/v2024.1/paper.pdf
