### 2024-07-08 #6864 @gomartini-collab
This is the kind of things I am not sure:

 **Data sharing**: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.

If I click here, I am accepting the paper has NO original data. This is how I understood this part. Now you asked me to click on it. I am very much confuse.
