### 2024-07-09 #6864 @srmnitc
@gomartini-collab Sorry for the confusion from my comments. I will try to clarify the procedure. If the paper had presented some results, for example from a simulation, this data should be available to you as a reviewer to verify.

If the paper does not present any original data/results calculated with the code, as you pointed out, there is no data itself for you to verify and hence this box should be checked, and marked as completed. Does this answer your question?

The reproducibility point would be similar. If there is no original data presented, you do not need to verify that it is reproducible. Therefore, once again, you can check, and mark it as completed.

'Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?', this should only be marked as completed if the code has automated or manual steps. If not, it would be great if you raise an issue on the repository, and mention what you think is lacking.

You can also see more explanations [here](https://joss.readthedocs.io/en/latest/review_criteria.html). Please let me know if its not clear, and thanks again for your time and effort in reviewing!
