### 2024-07-10 #6864 @junghans
@TariniHardikar thanks for your review!

Yes, we have around 300 tests, that are under the `test` target that `cmake` generates. We run them as part of the CI and during pull requests as well.

I could make that more clear in the installation guide if that helps.
