### 2024-07-10 c03f9f4
README: add ctest command
