+++ README.rst
@@ -20,6 +20,7 @@ Basic installation
     git clone -b ${version} https://github.com/votca/votca.git
     cmake -B builddir -S votca -DBUILD_XTP=ON -DCMAKE_INSTALL_PREFIX=${prefix}
     cmake --build builddir --parallel <number of cores>
+    ctest --test-dir builddir # to test build
     cmake --build builddir --target install
 
 Using this code via docker
