### 2024-07-10 bce2b88
INSTALL.rst: add crest documentation
