### 2024-07-10 #6864 @junghans
Documentation update in https://github.com/votca/votca/pull/1145.
