### 2024-07-10 #6864 @TariniHardikar
Thank you so much for clarifying, and adding the clear documentation, @junghans! Things look great to me, VOTCA is a very well documented and wonderful package!
