### 2024-07-12 927b986
JOSS paper: add README
