+++ papers/v2024.1/README.md
@@ -1,0 +1,3 @@
+VOTCA: multiscale frameworks for quantum and classical simulations in soft matter,
+B. Baumeier, J. Wehner, N. Renaud, F. Zapata Ruiz, R. Halver, P. Madhikar, R. Gerritsen, G. Tirimbo, J. Sijen, D. Rosenberger, J. S. Brown, V. Sundaram, J. Krajniak, M. Bernhardt, and C. Junghans,
+[J. Open Source Softw. (2024)](https://doi.org/10.21105/joss.06795).
