+++ README.rst
@@ -16,7 +16,7 @@ Basic installation
 ::
 
     prefix=WHERE/TO/INSTALL/VOTCA
-    version=master # or 'v2024'
+    version=master # or 'v2024.1'
     git clone -b ${version} https://github.com/votca/votca.git
     cmake -B builddir -S votca -DBUILD_XTP=ON -DCMAKE_INSTALL_PREFIX=${prefix}
     cmake --build builddir --parallel <number of cores>
