+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 For more detailed information about the changes see the history of the
 `repository <https://github.com/votca/votca/commits/master>`__.
 
+Version 2024.1-dev
+==================
+
 Version 2024.1 (released 16.07.24)
 ==================================
 
