### 2024-07-16 #6864 @junghans
> @junghans could you please take a look at the following items to finish this up:
>
> * [ ]  Double check authors and affiliations (including ORCIDs) - please check Javier Sijen
> * [ ]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [ ]  Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [ ]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [ ]  Make sure that the license listed for the archive is the same as the software license.
>
> If you cannot check the boxes here directly, feel free to copy it.

@srmnitc all done.
