+++ .github/workflows/freebsd.yml
@@ -3,8 +3,6 @@ on:
   push:
     branches:
       - master
-    tags:
-      - v[0-9]*
   pull_request:
     branches:
       - master
