### 2024-07-16 d58e8d9
Update CHANGELOG.rst
