### 2024-07-18 #6864 @srmnitc
@junghans I was taking a look at the paper. The paper describes the theory in some detail. This is not really a requirement for JOSS papers. My recommendation is that would be nice if you add the section from Coarse Graining until Code Structure into a documentation page, and link it in the main paper. Or you could a shorter theory part and still add the link.

what do you think?
