### 2024-07-18 #6864 @junghans
@srmnitc Those theory sections are already in the csg and xtp manuals on our [website](https://www.votca.org/), so the only part left would be to purge it from the paper. I, personally, like having the background right there in the paper to spare the reader from jumping to external links, but I am good either way.

@baumeier do you have an opinion?
