### 2024-07-18 #6864 @baumeier
@junghans @srmnitc My preference would be to leave it in, as I agree that jumping to external links is cumbersome. And I think it is nice to demonstrate in one place what the code actually implements. I don't see how we can shorten the theory part (at least the XTP-related one) as it is a compacted version already and any further shortening would make it not understandable.
