### 2024-07-18 #6864 @DRosen285
As one of the authors I agree with @junghans on having the background right there in the paper. This presents a concise way for users of VOTCA to see what is new in one place, instead of having to jump to multiple references to get the key information.
