### 2024-07-18 #6864 @srmnitc
@junghans @baumeier @DRosen285 thanks for giving your feedback. Lets keep the theory in. Then I only have some small suggestions to the paper, for which I will open a PR/issue later today. Thanks again for the feedback.
