### 2024-07-18 votca/votca#1147 @srmnitc
Changes for the v2024.1 JOSS paper
This is part of https://github.com/openjournals/joss-reviews/issues/6864

I suggest the following changes to the paper. Feel free to choose which points to implement as needed

- line 19: commas `electronic and, therefore, quantum`
- line 19 could be strengthened with a citation
- line 20 Object-**O**riented, it is a lower case 'o' in Fig. 1
- Fig 1 - typo in caption 'differnt'
- line 28: comma `in 2015,`
- line 29: commas `more examples, and`
- line 31, gained insight could use a citation
- line 35: Others have stopped....community. You might want to phrase it in a more neutral way, something like no longer maintained/or few developers. Just a suggestion
- line 36: to build **an** all-inclusive
- line 48: comma - `environment model, and to`
- line 56: comma - `iterative approach, the`
- line 60: hyphen between Monte-Carlo
- line 72: Is Tikhonov reg well known, or does it need a ref?
- line 89: upper case for Matrix
- line 109: maybe DFT abbreviation was not introduced?
- line 114: is italics needed for active
- line 122: QP was not introduced
- line 129: Bethe-Saltpeter was used before, maybe you could introduce abbreviation there
