### 2024-07-18 #6864 @srmnitc
@junghans I opened an issue with some suggestions for the paper. Please take a look and implement as you see fit.
