### 2024-07-19 ab48bbd
paper.md: implemented some of srmnitc's suggestions
