### 2024-07-19 votca/votca#1147 @DRosen285
This is a possible review for the Tikhonov regularization:
\bibitem{inverse_problems}
H. W. Engl, M. Hanke, A. Neubauer \textit{Regularization of Inverse Problems} (Kluver Academic Publishers, 2000).
