### 2024-07-19 34e58f2
Update paper.md
