### 2024-07-19 d287cc9
added citations
