+++ papers/v2024.1/paper.bib
@@ -1,16 +1,4 @@
-ceton University Press},
-   hon package for astronomy}",
-     journal = {Astronomy and Astrophysics},
-     archivePrefix = "arXiv",
-     eprint = {1307.6212},
-     primaryClass = "astro-ph.IM",
-     keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
-     year = 2013,
-     month = oct,
-     volume = 558,
-     doi = {10.1051/0004-6361/201322068},
-     url = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
- }
+
 
  @article{pyscf,
     author = {Sun, Qiming and Zhang, Xing and Banerjee, Samragni and Bao, Peng and Barbry, Marc and Blunt, Nick S. and Bogdanov, Nikolay A. and Booth, George H. and Chen, Jia and Cui, Zhi-Hao and Eriksen, Janus J. and Gao, Yang and Guo, Sheng and Hermann, Jan and Hermes, Matthew R. and Koh, Kevin and Koval, Peter and Lehtola, Susi and Li, Zhendong and Liu, Junzi and Mardirossian, Narbe and McClain, James D. and Motta, Mario and Mussard, Bastien and Pham, Hung Q. and Pulkin, Artem and Purwanto, Wirawan and Robinson, Paul J. and Ronca, Enrico and Sayfutyarova, Elvira R. and Scheurer, Maximilian and Schurkus, Henry F. and Smith, James E. T. and Sun, Chong and Sun, Shi-Ning and Upadhyay, Shiv and Wagner, Lucas K. and Wang, Xiao and White, Alec and Whitfield, James Daniel and Williamson, Mark J. and Wouters, Sebastian and Yang, Jun and Yu, Jason M. and Zhu, Tianyu and Berkelbach, Timothy C. and Sharma, Sandeep and Sokolov, Alexander Yu. and Chan, Garnet Kin-Lic},
@@ -27,6 +15,53 @@ ceton University Press},
     url = {https://doi.org/10.1063/5.0006074},
     eprint = {https://pubs.aip.org/aip/jcp/article-pdf/doi/10.1063/5.0006074/16722275/024109\_1\_online.pdf},
 }
+
+
+@article{Lee2019,
+author = {Lee, Changyeon and Lee, Seungjin and Kim, Geon-U and Lee, Wonho and Kim, Bumjoon J.},
+title = {Recent Advances, Design Guidelines, and Prospects of All-Polymer Solar Cells},
+journal = {Chemical Reviews},
+volume = {119},
+number = {13},
+pages = {8028-8086},
+year = {2019},
+doi = {10.1021/acs.chemrev.9b00044},
+    note ={PMID: 31181904},
+URL = {https://doi.org/10.1021/acs.chemrev.9b00044},
+eprint = {https://doi.org/10.1021/acs.chemrev.9b00044}
+}
+
+@article{Svaneborg2023,
+    author = {Svaneborg, Carsten and Everaers, Ralf},
+    title = "{Multiscale equilibration of highly entangled isotropic model polymer melts}",
+    journal = {The Journal of Chemical Physics},
+    volume = {158},
+    number = {5},
+    pages = {054903},
+    year = {2023},
+    month = {02},
[…]
