### 2024-07-19 9278338
Update CHANGELOG.rst
