+++ papers/v2024.1/paper.bib
@@ -52,16 +52,19 @@ eprint = {https://doi.org/10.1021/acs.chemrev.9b00044}
 
 
 
-@book{engl2000regularization,
+@Inbook{engl2000regularization,
   title={Regularization of Inverse Problems},
-  author={Engl, H.W. and Hanke, M. and Neubauer, A.},
-  isbn={9780792361404},
-  lccn={96028672},
-  series={Mathematics and Its Applications},
-  url={https://books.google.nl/books?id=VuEV-Gj1GZcC},
-  year={2000},
-  publisher={Springer Netherlands}
-} 
+  author={Engl, Heinz W. and Ramlau, Ronny},
+  editor={Engquist, Bj{\"o}rn},
+  bookTitle={Encyclopedia of Applied and Computational Mathematics},
+  year={2015},
+  publisher={Springer Berlin Heidelberg},
+  address={Berlin, Heidelberg},
+  pages={1233--1241},
+  isbn={978-3-540-70529-1},
+  doi={10.1007/978-3-540-70529-1_52},
+  url={https://doi.org/10.1007/978-3-540-70529-1_52}
+}
 
 
 @article{turbomole,
