+++ papers/v2024.1/paper.bib
@@ -48,6 +48,19 @@ eprint = {https://doi.org/10.1021/acs.chemrev.9b00044}
 }
 
 
+@article{Rudzinski_2021,
+  doi = {10.1088/1361-648X/abe6e1},
+  url = {https://dx.doi.org/10.1088/1361-648X/abe6e1},
+  year = {2021},
+  month = {may},
+  publisher = {IOP Publishing},
+  volume = {33},
+  number = {22},
+  pages = {224001},
+  author = {Joseph F Rudzinski and Sebastian Kloth and Svenja Wörner and Tamisra Pal and Kurt Kremer and Tristan Bereau and Michael Vogel},
+  title = {Dynamical properties across different coarse-grained models for ionic liquids},
+  journal = {Journal of Physics: Condensed Matter}
+}
 
 
 
