### 2024-07-20 942ea99
scripts: fix exit code in tutorials
