+++ csg-tutorials/propane/imc-ibiintra/pre/run.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/usr/bin/env -S bash -e
 set -e
 
 # pre-iterations with IBI
