### 2024-07-22 #6864 @junghans
Proof looks good! @baumeier @DRosen285 please check as well.
2 nitpicks for citations, twice it says "A. Lyubartsec et. al" instead of just the last name like for all the other references. And "Andrienko & others" should be "Andrienko et. al".

For the former issue I assume the JOSS converter does not like `author = {Last, First}`, but only support author={First Last}` (CC @kyleniemeyer)
