### 2024-07-22 eb85084
paper.bib: update ref
