+++ papers/v2024.1/paper.bib
@@ -485,7 +485,7 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 
 @software{mpip-votca,
 	title = {MPI-P Fork of VOTCA},
-        author = {Andrienko, Denis and {others}},
+        author = {Andrienko, Denis and Scherer, Christoph and Stanisic, Luka and {others}},
         version = {2023.1},
         date = {2023-12-07},
 	url = {https://gitlab.mpcdf.mpg.de/votca/votca}
@@ -501,7 +501,7 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 	year = {2020},
 	doi = {10.1021/acs.jctc.9b01256},
 	note ={PMID: 32282206},
-	URL = { ttps://doi.org/10.1021/acs.jctc.9b01256 },
+	URL = { https://doi.org/10.1021/acs.jctc.9b01256 },
 	eprint = { https://doi.org/10.1021/acs.jctc.9b01256 }
 }
 
