+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ For more detailed information about the changes see the history of the
 Version 2024.1-dev
 ==================
 
--  Editorial changes for v2024.1 JOSS paper  (#1148)
+-  Editorial changes for v2024.1 JOSS paper  (#1148, #1153)
 -  Fix permissions on some scripts (#1150)
 -  Fix exit code on some scripts (#1151)
 
