### 2024-07-22 10d403a
paper.bib: add Alex's middle
