+++ papers/v2024.1/paper.bib
@@ -286,7 +286,7 @@ year = {2017}
 	language = {en},
 	urldate = {2024-02-20},
 	journal = {Faraday Discuss.},
-	author = {Lyubartsev, Alexander and Mirzoev, Alexander and Chen, LiJun and Laaksonen, Aatto},
+	author = {Lyubartsev, Alexander P. and Mirzoev, Alexander and Chen, LiJun and Laaksonen, Aatto},
 	year = {2010},
 	pages = {43--56},
 	file = {Eingereichte Version:/Users/davidrosenberger/Zotero/storage/ZXJ4DZPK/Lyubartsev et al. - 2010 - Systematic coarse-graining of molecular models by .pdf:application/pdf},
