### 2024-07-22 #6864 @junghans
> @junghans I believe the paper converter will add first initials when needing to distinguish multiple citations from the same first author; it should be able to handle the `author = {Last, First and Last2, First2}` format just fine.

@kyleniemeyer thanks. In our case there is no distinguish needed, I think we are missing a middle initial somewhere, so let me see if that fixes it.
