### 2024-07-22 #6864 @junghans
Ok fixed in https://github.com/votca/votca/pull/1153
