### 2024-07-23 fb04d81
Update CHANGELOG.rst
