### 2024-07-23 b2a4eaf
Update CITATION.cff
