### 2024-07-23 9941021
Update CITATION.cff
