### 2024-07-23 04771d5
add some affiliations back
