### 2024-07-23 c0f4e13
Update CITATION.cff
