### 2024-07-23 966e7aa
Update CHANGELOG.rst
