+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Version 2024.1-dev
 -  Editorial changes for v2024.1 JOSS paper  (#1148, #1153, #1154)
 -  Fix permissions on some scripts (#1150)
 -  Fix exit code on some scripts (#1151)
+-  Update CITATION.cff (#1155)
 
 Version 2024.1 (released 16.07.24)
 ==================================
